refactor(web): split the Postgres reader out of projection/contract.ts
Addresses audit finding 8: contract.ts was importing the pg driver
directly, which meant any browser-reachable surface that took a runtime
value from the contract risked dragging the driver toward the client
bundle. snapshot-reader.ts is now the only importer of pg in the
projection module; asOfSortKey moved above the seam into as-of.ts,
whose header now explains the reason that actually holds now that the
contract is pg-free rather than the original slice-#150 rationale. The
re-point set is dashboard.ts (imports getReaderPool/getSnapshotHistory
from the new reader module) plus the split tests: contract.test.ts
sheds its reader-backed cases (moved into the new
snapshot-reader.test.ts) and client-bundle.integration.test.ts updates
its forbidden-string commentary to point at the new module.
